Move singleton to utils (#340)
