+++ towhee/utils/singleton.py
